reorg(expansions) split out offsetFromPoints offsetGeom2 offsetPath2 adjusted code and test suites
