+++ src/operations/expansions/expandPath2.js
@@ -2,7 +2,7 @@ const { vec2 } = require('../../math')
 
 const { geom2, path2 } = require('../../geometry')
 
-const { offsetFromPoints } = require('./offset')
+const offsetFromPoints = require('./offsetFromPoints')
 
 const area = require('./area')
