+++ src/operations/expansions/index.js
@@ -1,4 +1,4 @@
 module.exports = {
   expand: require('./expand'),
-  offset: require('./offset').offset
+  offset: require('./offset')
 }
